requirements.txt
Flask~=3.1.2
PyJWT~=2.10.1
redis~=6.4.0
Booktype~=1.5
dotenv~=0.9.9
python-dotenv~=1.1.1
flasgger~=0.9.7.1
psycopg2~=2.9.10
pyotp~=2.9.0
bcrypt~=4.3.0